test: restructure suite and load plugins earlier
