--- plugin/tests/test-sample.php
@@ -1,20 +1,0 @@
-<?php
-/**
- * Class SampleTest
- *
- * @package lhpbp\plugin
- */
-
-/**
- * Sample test case.
- */
-class SampleTest extends WP_UnitTestCase {
-
-	/**
-	 * A single example test.
-	 */
-	public function test_sample() {
-		// Replace this with some actual testing code.
-		$this->assertTrue( true );
-	}
-}
