+++ tests/plugin/Lhpbpp_Test.php
